requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
rdkit>=2022.9.0
tape-proteins==0.5
hydra-core>=1.3.0
omegaconf>=2.3.0
lifelines>=0.27.0
tensorboard>=2.14.0
tensorboardX>=2.6.0
tqdm>=4.60.0
matplotlib>=3.7.0
biopython>=1.83
PyYAML>=6.0
transformers>=4.30.0

# Optional: Distributed training
# horovod>=0.28.0
